test: use 1 file for presidio guard tests
